+++ number_insight/src/vonage_number_insight/__init__.py
@@ -1,3 +1,35 @@
+from . import errors
 from .number_insight import NumberInsight
+from .requests import (
+    AdvancedAsyncInsightRequest,
+    AdvancedSyncInsightRequest,
+    BasicInsightRequest,
+    StandardInsightRequest,
+)
+from .responses import (
+    AdvancedAsyncInsightResponse,
+    AdvancedSyncInsightResponse,
+    BasicInsightResponse,
+    CallerIdentity,
+    Carrier,
+    RealTimeData,
+    RoamingStatus,
+    StandardInsightResponse,
+)
 
-__all__ = ['NumberInsight']
+__all__ = [
+    'NumberInsight',
+    'BasicInsightRequest',
+    'StandardInsightRequest',
+    'AdvancedAsyncInsightRequest',
+    'AdvancedSyncInsightRequest',
+    'BasicInsightResponse',
+    'CallerIdentity',
+    'Carrier',
+    'RealTimeData',
+    'RoamingStatus',
+    'StandardInsightResponse',
+    'AdvancedSyncInsightResponse',
+    'AdvancedAsyncInsightResponse',
+    'errors',
+]