+++ number_insight/src/vonage_number_insight/responses.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import Literal, Optional, Union
 
 from pydantic import BaseModel
 
@@ -15,13 +15,61 @@ class BasicInsightResponse(BaseModel):
     country_prefix: Optional[str] = None
 
 
+class Carrier(BaseModel):
+    network_code: Optional[str] = None
+    name: Optional[str] = None
+    country: Optional[str] = None
+    network_type: Optional[str] = None
+
+
+class CallerIdentity(BaseModel):
+    caller_type: Optional[str] = None
+    caller_name: Optional[str] = None
+    first_name: Optional[str] = None
+    last_name: Optional[str] = None
+    subscription_type: Optional[str] = None
+
+
 class StandardInsightResponse(BasicInsightResponse):
-    ...
+    request_price: Optional[str] = None
+    refund_price: Optional[str] = None
+    remaining_balance: Optional[str] = None
+    current_carrier: Optional[Carrier] = None
+    original_carrier: Optional[Carrier] = None
+    ported: Optional[str] = None
+    caller_identity: Optional[CallerIdentity] = None
+    caller_type: Optional[str] = None
+    caller_name: Optional[str] = None
+    first_name: Optional[str] = None
+    last_name: Optional[str] = None
+
+
+class RoamingStatus(BaseModel):
+    status: Optional[str] = None
+    roaming_country_code: Optional[str] = None
+    roaming_network_code: Optional[str] = None
+    roaming_network_name: Optional[str] = None
+
+
+class RealTimeData(BaseModel):
+    active_status: Optional[str] = None
+    handset_status: Optional[str] = None
+
+
+class AdvancedSyncInsightResponse(StandardInsightResponse):
+    roaming: Optional[Union[RoamingStatus, Literal['unknown']]] = None
+    lookup_outcome: Optional[int] = None
+    lookup_outcome_message: Optional[str] = None
+    valid_number: Optional[str] = None
+    reachable: Optional[str] = None
+    real_time_data: Optional[RealTimeData] = None
+    ip_warnings: Optional[str] = None
 
 
 class AdvancedAsyncInsightResponse(BaseModel):
-    ...
-
-
-class AdvancedSyncInsightResponse(BaseModel):
-    ...
+    request_id: Optional[str] = None
+    number: Optional[str] = None
+    remaining_balance: Optional[str] = None
+    request_price: Optional[str] = None
+    status: Optional[int] = None
+    error_text: Optional[str] = None