+++ vonage/CHANGES.md
@@ -1,3 +1,6 @@
+# 3.99.0a8
+- Add support for the [Vonage Number Insight API](https://developer.vonage.com/en/number-insight/overview).
+
 # 3.99.0a7
 - Add support for the [Vonage Voice API](https://developer.vonage.com/en/voice/voice-api/overview).
 - Add `http_client` property to each module that has an HTTP Client, e.g. Voice, Sms, Verify.